Solution: Day 3
Fun one, but I should have realized sooner that the maximum number was
just the max digit with the lowest index. Tried to solve using
combinations of 12, obviously that wouldnt work but I just wanted to
see how the performance was. I also tried the inverse of removing
digits since it was only 3 of them in the example input but I should
have realized there was a lot more in the actual input sooner. Once the
property clicked it was easy enough and of course much more performant.