Solution: Day 5
The second part was a nice one, took me a while to realize the fusing
process of adding a new range to the regions. You probably can scan it
if the data was sorted, I'll have to look into this, but my brute force
approach was just to try to fuse a new region with the first valid one
and if they fuse, remove the old and just try to insert the fused one.
It's still fast (~4ms), but we could do better!