Solution: Day 6
First part was easy as allways, but I got tricked by the example input
being a constant width for every sum/product and assumed that when doing
part 2. Once it finally clicked I then made the second blunder of trying
to solve it without transposing which made it harder than it is.

The version I upload now is a revised one which transposes everything
and is in my opionon cleaner than the first part.