Solution: Day 8
Finally the difficulty is ramping up! I know the intended algorithm is
union find, I just didn't know it at the top of my head so I had to
implement the brute force logic instead. This is no issue however, as
the most computationally demanding part is calculating the distances.
I'll try a rust version as well, but I highly doubt using the fancier
algorithm will give much of a perfomance boost.