Solution: Day 10
The number of lines in the final solution does not do the amount of work
I spent on this justice.

This was hard, I tried for a long time to do a
search like part 1, with larger steps and other tricks. Not feasable.

I then tried to solve it both symbolically and then numericallyy, but
this was obviously not doable as the buttons are not linearly
independant.

I then attempted Linear Least Square, but since the solution requires
integers this was doomed to fail. I nevertheless spent a lot of time
tryng to implement a solution which used the LLS result to backstep and
in that way correct the solution to an integer one. This did not work
either as it required a lot of backstepping (more than I was able to
compute).

In the end I just stumbled on the mixed-integer linear
programming optimization method in scipy and I was finally able to
deduce how to convert the problem into that domain.

I would have needed to spend a huge amount of time rolling my own
implementation of this, so like yesterday, I'm breaking my rule and
using libraries to solve these problems.

However, I did learn something at least, so that's good!